Only use fake mutex during parallel tests
Travis doesn't like these filters

+++ features/support/quarantine.rb
@@ -5,13 +5,15 @@ Around("@quarantine-files") do |scenario, block|
 end
 
 # Simulate a mutex to stop these scenarios stomping on each other during parallel test runs
-Before("@quarantine-files") do
-  while File.exists?("tmp/cucumber_quarantine_files")
-    sleep(0.2)
+if ENV['TEST_ENV_NUMBER']
+  Before("@quarantine-files") do
+    while File.exists?("tmp/cucumber_quarantine_files")
+      sleep(0.2)
+    end
+    File.open("tmp/cucumber_quarantine_files", "w") {}
   end
-  File.open("tmp/cucumber_quarantine_files", "w") {}
-end
 
-After("@quarantine-files") do
-  File.delete("tmp/cucumber_quarantine_files")
+  After("@quarantine-files") do
+    File.delete("tmp/cucumber_quarantine_files")
+  end
 end
